Make ToastContainer its own separate module
